Standalone psp->falco rules converter in go
Break out the PSP to falco rules converter to a standalone program,
written in go. It takes a PSP as input and writes a falco rules file as
output. It uses golang templating to render the set of rules and uses
k8s PodSecurityPolicy objects to hold the PSP after being read.

This version passes 71/74 of the automated tests.