Don't stop priv. images w/ PSP AllowPrivEscalation
Previously if a PSP had AllowPrivilegeEscalation set to false, it would
also detect privileged images.

A stricter interpretation of the docs imply that the
AllowPrivilegeEscalation option should *only* affect whether someone
creates a container with the AllowPrivilegeEscalation option set in the
security context, so change the tests and behavior to match that more
strict interpretation.